+++ crates/interpreter/src/error.rs
@@ -60,7 +60,7 @@ pub enum Unsupported {
 pub fn print_backtrace() {
     let backtrace = std::backtrace::Backtrace::capture();
     if matches!(backtrace.status(), std::backtrace::BacktraceStatus::Captured) {
-        println!("{}", backtrace);
+        println!("{backtrace}");
     }
 }