+++ crates/scheduler/src/protocol.rs
@@ -226,7 +226,9 @@ fn process_install<B: Board>(
             board::Applet::<B>::finish()?;
             *scheduler.protocol.0.install() = TransferState::Ready;
             if !dry_run {
-                scheduler.start_applet().inspect_err(|_| log::warn!("Failed to start applet."))?;
+                scheduler
+                    .start_applet()
+                    .inspect_err(|e| log::warn!("Failed to start applet: {}", e))?;
             }
         }
     }