+++ crates/store/src/buffer.rs
@@ -309,7 +309,7 @@ impl Storage for BufferStorage {
         self.options.max_page_erases
     }
 
-    fn read_slice(&self, index: StorageIndex, length: usize) -> Result<Cow<[u8]>, Error> {
+    fn read_slice(&self, index: StorageIndex, length: usize) -> Result<Cow<'_, [u8]>, Error> {
         Ok(Cow::Borrowed(&self.storage[index.range(length, self)?]))
     }