+++ scripts/sync.sh
@@ -36,6 +36,7 @@ for dir in $(find crates examples/rust -name Cargo.toml -printf '%h\n' | sort); 
   # TODO(https://github.com/rust-lang/rust-clippy/issues/13994): Remove when fixed.
   add_lint $file allow clippy.literal-string-with-formatting-args
   add_lint $file warn clippy.mod-module-files
+  [ $crate = $dir ] || add_lint $file allow clippy.uninlined_format_args
   add_lint $file allow clippy.unit-arg
   # add_lint $file warn rust.elided-lifetimes-in-paths
   # TODO: Use the same [ -e src/lib.rs -a "$(package_publish)" = true ] test as in test-helper.