+++ scripts/wrapper.sh
@@ -57,18 +57,18 @@ case "$1" in
   cargo)
     case "$2" in
       bloat) ensure_cargo cargo-bloat 0.12.1 ;;
-      upgrade) ensure_cargo cargo-edit 0.13.1 ;;
+      upgrade) ensure_cargo cargo-edit 0.13.6 ;;
       *) e "Wrapper does not support 'cargo $2'" ;;
     esac
     ;;
   defmt-print) ensure_cargo defmt-print 1.0.0-rc.1 ;;
-  mdbook) ensure_cargo mdbook 0.4.47 ;;
-  nrfdfu) ensure_cargo nrfdfu 0.1.3 ;;
-  probe-rs) ensure_cargo probe-rs-tools 0.27.0 ;;
+  mdbook) ensure_cargo mdbook 0.4.51 ;;
+  nrfdfu) ensure_cargo nrfdfu 0.2.0 ;;
+  probe-rs) ensure_cargo probe-rs-tools 0.29.0 ;;
   rust-objcopy|rust-size) ensure_cargo cargo-binutils 0.3.6 ;;
-  taplo) ensure_cargo taplo-cli 0.9.3 ;;
-  trunk) ensure_cargo trunk 0.21.8 ;;
-  twiggy) ensure_cargo twiggy 0.7.0 ;;
+  taplo) ensure_cargo taplo-cli 0.10.0 ;;
+  trunk) ensure_cargo trunk 0.21.14 ;;
+  twiggy) ensure_cargo twiggy 0.8.0 ;;
   *) IS_CARGO=n ;;
 esac
 [ $IS_CARGO = y ] && run "$@"