+++ crates/cli-tools/CHANGELOG.md
@@ -26,7 +26,7 @@
 
 ### Patch
 
-- Fix clippy lints
+- Fix lints
 - Update dependencies
 
 ## 0.2.0
@@ -78,4 +78,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 9 -->
+<!-- Increment to skip CHANGELOG.md test: 10 -->