+++ crates/scheduler/CHANGELOG.md
@@ -17,7 +17,7 @@
 
 - Use `cargo::` instead of `cargo:` for build script instructions
 - Add warning when applet fails to start after install
-- Fix clippy lints
+- Fix lints
 - Update dependencies
 
 ## 0.4.0
@@ -191,4 +191,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 8 -->
+<!-- Increment to skip CHANGELOG.md test: 9 -->